requirements.txt
llama-index
langchain
ollama
faiss-cpu
python-dotenv
streamlit
sqlalchemy
tqdm
requests
sentence-transformers